requirements.txt
telethon>=1.34.0
python-telegram-bot>=20.7
pytesseract>=0.3.10
pillow>=10.1.0
aiohttp>=3.9.1
python-dotenv>=1.0.0
loguru>=0.7.2
opencv-python>=4.9.0
paddleocr>=2.7.3
transformers>=4.36.2
requests>=2.31.0
torch>=2.1.2
SQLAlchemy==2.0.30
